% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2pMax.R
\name{d2pMax}
\alias{d2pMax}
\title{d2pMax : provides pMax given dMax and nVar}
\usage{
d2pMax(nVar, dMaxKnown)
}
\arguments{
\item{nVar}{The model dimension expected. This parameter will be
deduced from the input data (\code{series}) if \code{series} is
a matrix. If \code{series} is a vector, the expected dimension nVar
should be provided.}

\item{dMaxKnown}{The number of polynoms terms to retrieve}
}
\value{
The maximum polynomial degree \code{dMax} used to code the polynomial
}
\description{
Gives the maximum polynomial degree pMax
             given the number of polynomial terms dMax
             and the system dimension nVar
}
\examples{
#############
# Example 1 #
#############
# Maximum polynomial degree ?
# number of variables:
nVar <- 3
# polynomial degree:
dMax <- 3
# The maximal polynomial degree used for coding the polynomial is:
d2pMax(nVar,dMax)

}
\author{
Sylvain Mangiarotti
}
\seealso{
\code{\link{gloMoId}}, \code{\link{gPoMo}}, \code{\link{poLabs}}
}

