% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vignetteIII-data.R
\docType{data}
\name{data_vignetteIII data set}
\alias{data_vignetteIII data set}
\alias{data_vignetteIII}
\title{Output of the vignette \code{III_Modelling}}
\format{
An object of class \code{list} of length 12.
}
\usage{
data_vignetteIII
}
\description{
To reduce the computation time, the outputs
of the simulations presented in vignette VI have been
run beforehand and saved in this file.
}
\author{
Sylvain Mangiarotti, Mireille Huc.
}
\keyword{data}
