% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vignetteVI-data.R
\docType{data}
\name{data_vignetteVI data set}
\alias{data_vignetteVI data set}
\alias{data_vignetteVI}
\title{Output of the vignette \code{VI_Sensitivity}}
\format{
An object of class \code{list} of length 6.
}
\usage{
data_vignetteVI
}
\description{
To reduce the computation time, the outputs
of the simulations presented in vignette VI have been
run beforehand and saved in this file.
}
\author{
Sylvain Mangiarotti, Mireille Huc.
}
\keyword{data}
