% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vignetteVII-data.R
\docType{data}
\name{data_vignetteVII data set}
\alias{data_vignetteVII data set}
\alias{data_vignetteVII}
\title{Output of the vignette \code{VII_Retro-Modelling}}
\format{
An object of class \code{list} of length 29.
}
\usage{
data_vignetteVII
}
\description{
To reduce the computation time, the outputs
of the simulations presented in vignette VII have been
run beforehand and saved in this file.
}
\author{
Sylvain Mangiarotti, Mireille Huc.
}
\keyword{data}
