% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectP1limCycl.R
\name{detectP1limCycl}
\alias{detectP1limCycl}
\title{detectP1limCycl : Detect limit cycles of period-1}
\usage{
detectP1limCycl(data, LimCyclThreshold = 0.01, show = 2)
}
\arguments{
\item{data}{A matrix of the trajectory in a 2D space (only
the two first column are considered)}

\item{LimCyclThreshold}{The detection threshold}

\item{show}{Indicates the deepness of the feedback (from 0 to 2)}
}
\value{
indicates if a limit cycle is detected (1) or not (0)
}
\description{
Detects the limit cycle from trajectories
in a bidimensional projection
}
\author{
Sylvain Mangiarotti
}

