% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEq.R
\name{extractEq}
\alias{extractEq}
\title{extractEq : Extracts Equations from one system}
\usage{
extractEq(KL, eqVect)
}
\arguments{
\item{KL}{A model, provided as a matrix.}

\item{eqVect}{A vector of integers, providing the equations
numbers to be kept in the output equation system.}
}
\description{
Combines equations of different sources
into a single system. During this combination, the polynomial
maximal degree can be either imposed or optimized to reduce
the model size. All the input have to follow
the convention defined by \code{poLabs}.
}
\author{
Mireille Huc
}
