% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P1FxCh-data.R
\docType{data}
\name{P1FxCh}
\alias{P1FxCh}
\title{A data set for testing periodicity}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 6 columns.
}
\usage{
P1FxCh
}
\description{
A matrix of 6 columns corresponding to six
time series, two resulting from a Period-1 limit cycle,
two from regime converging to fixed point, and two relating
to a chaotic behavior
}
\author{
Sylvain Mangiarotti, Mireille Huc.
}
\keyword{data}
