% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P1FxChP2-data.R
\docType{data}
\name{P1FxChP2}
\alias{P1FxChP2}
\title{A data set for testing periodicity}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 8 columns.
}
\usage{
P1FxChP2
}
\description{
Trajectories for testing periodicity.
The following regimes are made available:
Period-1 in columns 1:2, Fixed Point in 3:4,
chaotic in 5:6, Period-2 in 7:8
}
\author{
Sylvain Mangiarotti, Mireille Huc.
}
\keyword{data}
