% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramId.R
\name{paramId}
\alias{paramId}
\title{paramId : parameter Identification}
\usage{
paramId(allForK, drv, weight)
}
\arguments{
\item{allForK}{The list of input parameters}

\item{drv}{the derivative (on the equation left hand)}

\item{weight}{The weighting series}
}
\value{
allForK The initial list completed with the model parameters.
}
\description{
Estimate polynomial coefficients.
}
\author{
Sylvain Mangiarotti
}

