% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svrlTS-data.R
\docType{data}
\name{svrlTS}
\alias{svrlTS}
\title{A data set for the global modeling of time series in association}
\format{An object of class \code{list} of length 4.}
\usage{
svrlTS
}
\description{
This data set aims to test the global modelling technique
when several time series of different sizes are available.
Four time series are provided, all derived from the Rössler-1976 system.
}
\references{
S. Mangiarotti, F. Le Jean, M. Huc & C. Letellier, 2016.
Global modeling of aggregated and associated chaotic dynamics,
Chaos, Solitons & Fractals, 83, 82-96.
}
\author{
Sylvain Mangiarotti, Mireille Huc.
}
\keyword{data}
