% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS-data.R
\docType{data}
\encoding{UTF-8}
\name{TS}
\alias{TS}
\title{Time series resulting from the integration of a non stationary system}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6001 rows and 2 columns.
}
\usage{
TS
}
\description{
A 2*6001 matrix with the time vector in column one
and a time series resulting from the integration of a non stationary
Rössler system -- parameter a varying in time: a(t) -- in colmn two.
}
\author{
Sylvain Mangiarotti, Mireille Huc.
}
\keyword{data}
