% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wInProd.R
\name{wInProd}
\alias{wInProd}
\title{Weighted inner product}
\usage{
wInProd(A1, A2, weight = NULL)
}
\arguments{
\item{A1}{The input matrix 1.}

\item{A2}{The input matrix 2.}

\item{weight}{The weighting vector.}
}
\value{
\code{inP} The weighted inner product.
}
\description{
Computes weighted inner products.
}
\examples{
############
#Example 1 #
############
A1 = c(0,1,2,0,1,3)
A2 = c(1,2,0,0,4,1)
wInProd(A1, A2)

############
#Example 2 #
############
A1 = c(0,1,2,0,1,3)
A2 = c(1,2,0,0,4,1)
w = c(0,0,0,1,1,1)
wInProd(A1, A2, weight = w)

}
