% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivODE2.R
\name{derivODE2}
\alias{derivODE2}
\title{deriveODE2 : A Subfonction for the numerical integration
of polynomial equations in the generic form defined by function
\code{poLabs}}
\usage{
derivODE2(t, x, K, regS = NULL)
}
\arguments{
\item{t}{All the dates for which the result of the numerical integration
of the model will have to be provided}

\item{x}{Current state vector (input from which the next state will
be estimated)}

\item{K}{is the model: each column corresponds to one equation
which organisation is following the convention given by function
\code{poLabs} which requires the definition of the model dimension
\code{nVar} (i.e. the number of variables) and the maximum polynomial
degree \code{dMax} allowed.}

\item{regS}{Current states of each polynomial terms used
in \code{poLabs}. These states can be deduced from the current
state vector x (using function \code{regSeries}). When available,
it can be provided as an input to avoid unecessary computation.}
}
\description{
This function provides the one step integration of
polynomial Ordinary Differential Equations (ODE). This function
requires the function \code{ode} ("deSolve" package).
}
\author{
Sylvain Mangiarotti

#@export
}

