% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDVI-data.R
\docType{data}
\name{NDVI}
\alias{NDVI}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 9618 rows and 4 columns.}
\usage{
NDVI
}
\description{
A time series of 28 years of Normalized
Difference Vegetation Index measured from space based
on the Advanced Very High Resolution Radiometer (AVHRR)
sensor from 1982 to 2008 (see reference [1] for details).
}
\author{
Sylvain Mangiarotti, Flavie Le Jean
}
\references{
[1] Mangiarotti S., Drapeau L. & Letellier C., 2014.
Two chaotic models for cereal crops observed from satellite in 
northern Morocco.\cr
}
\keyword{data}

