% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosYco-data.R
\docType{data}
\name{RosYco}
\alias{RosYco}
\title{Twelve Rossler-1976 time series (exclusively variable \eqn{y})}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3000 rows and 12 columns.
}
\usage{
RosYco
}
\description{
Twelve independant Rossler-1976 time series
(variable y). The parameters used to generate the time
series correspond to a phase coherent behavior.
Details can be found in [1]
}
\details{
Another set of time series of the Rossler-1976 chaotic system
}
\references{
[1] Mangiarotti S., Le Jean F., Huc M. & Letellier C.,
Global Modeling of aggregated and associated chaotic
dynamics. Chaos, Solitons and Fractals, 83, 82-96.
}
\author{
Sylvain Mangiarotti, Flavie Le Jean.
}
\keyword{data}
